pyproject.toml
[project]
name = "grocerly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
    "django-types>=0.22.0",
]
